'''@ Reply 1.6.5''': fixed bug with _install.php file (closes #671)
git-svn-id: https://svn.dotclear.net/lab@3055 b35155ea-a742-4b26-8089-0891bb790021

+++ plugins/atReply/_install.php
@@ -2,7 +2,7 @@
 # ***** BEGIN LICENSE BLOCK *****
 #
 # This file is part of @ Reply, a plugin for Dotclear 2
-# Copyright (C) 2008,2009,2010 Moe (http://gniark.net/) and buns
+# Copyright (C) 2008,2009,2010,2011 Moe (http://gniark.net/) and buns
 #
 # @ Reply is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License v2.0
@@ -30,13 +30,23 @@ if (!defined('DC_CONTEXT_ADMIN')) {return;}
 # On lit la version du plugin
 $m_version = $core->plugins->moduleInfo('atReply','version');
 
+# On lit la version du plugin dans la table des versions
+$i_version = $core->getVersion('atReply');
+ 
+# La version dans la table est supérieure ou égale à
+# celle du module, on ne fait rien puisque celui-ci
+# est installé
+if (version_compare($i_version,$m_version,'>=')) {
+	return;
+}
+
 # delete this setting which was saved in the wrong namespace
 $core->con->execute('DELETE FROM '.$core->prefix.'setting '.
 		'WHERE (setting_ns = \'atreply\') '.
 		'AND (setting_id = \'wiki_comments\');');
 
 # La procédure d'installation commence vraiment là
-$core->setVersion('subscribeToComments',$m_version);
+$core->setVersion('atReply',$m_version);
 
 return true;
 ?>
